requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
more-itertools==8.3.0
numpy==1.18.4
packaging==20.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.17
threadpoolctl==2.0.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
mock 